% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listarCargosOrgaosLegislativosCD.R
\name{listarCargosOrgaosLegislativosCD}
\alias{listarCargosOrgaosLegislativosCD}
\title{List Legislative's Positions at the Camara dos Deputados}
\usage{
listarCargosOrgaosLegislativosCD()
}
\value{
A data frame with information on positions at the Camara dos Deputados
}
\description{
Returns a data frame that lists positions (for example, "Presidente", "Relator",
etc) that can be occupied by legislators in internal Camara dos Deputados organizations
(committees, floor, etc) and respective identification codes at the web service. This function
does not require any parameters.
}
\note{
The output of this function can be used as a parameter in functions that require
the type of propositions.
}
\examples{

listarCargosOrgaosLegislativosCD()

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

