% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listarPresencasDia.R
\name{listarPresencasDia}
\alias{listarPresencasDia}
\title{List Legislators Attendance in one day}
\usage{
listarPresencasDia(data, numLegislatura = "", numMatriculaParlamentar = "",
  siglaPartido = "", siglaUF = "")
}
\arguments{
\item{data}{string of format dd/mm/yyyy, the date if the sessions.}

\item{numLegislatura}{integer, the number specific legislature.}

\item{numMatriculaParlamentar}{integer, the legislator "matricula" number (check obterDeputados
function for help). This is a optional parameter and it's default is empty.}

\item{siglaPartido}{string, the political part of the proposition's author. This is a optional
parameter and it's default is empty.}

\item{siglaUF}{string, the state (Unidade da Federação) of the proposition's author. This is
a optional parameter and it's default is empty.}
}
\value{
A data frame with information on legislators attendance to sessions in an specific day.
}
\description{
Returns a data frame with information on legislators attendance to sessions
in an specific day. The maximum difference between initial and final date allowed by the
web service allowed is 7 days. The inputs for this functions are of class character and
in the format dd/mm/yyyy.
}
\examples{
dataSessao = "10/04/2012"
presencas <- listarPresencasDia(dataSessao)
head(presencas)
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

