% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listarProposicoes.R
\name{listarProposicoes}
\alias{listarProposicoes}
\title{List Proposals}
\usage{
listarProposicoes(sigla = "", ano = "", numero = "",
  datApresentacaoIni = "", datApresentacaoFim = "", idTipoAutor = "",
  parteNomeAutor = "", siglaPartidoAutor = "", siglaUfAutor = "",
  generoAutor = "", idSituacaoProposicao = "",
  idOrgaoSituacaoProposicao = "", emTramitacao = "")
}
\arguments{
\item{sigla}{string, the type of the proposal(s) (check listarSiglasTipoProposicao
function for help), which is part of the name of the proposal(s).}

\item{ano}{integer, the year of the proposal(s) (check listarSiglasTipoProposicao
function for help), which is part of the name of the proposal(s) and represents the year
the proposal was written.}

\item{numero}{integer, the number of the proposal(s) (check listarSiglasTipoProposicao function
for help), which is part of the name of the proposal(s). This is an optional parameter and
its default is empty.}

\item{datApresentacaoIni}{string of format dd/mm/yyyy, the initial date of the requested
proposals. This is an optional parameter and its default is empty.}

\item{datApresentacaoFim}{string of format dd/mm/yyyy, the initial date of for the requested
proposals. This is an optional parameter and its default is empty.}

\item{idTipoAutor}{character, the id code of the type of author of proposal (for example, "Bancada",
"Deputado Federal", "Senador", "Comissao Diretora", "Orgao do Poder Executivo", etc - check
listarTiposAutores function for help). This is an optional parameter and its default is empty.}

\item{parteNomeAutor}{string, a part of the name of the proposal author. This is an optional
parameter and its default is empty.}

\item{siglaPartidoAutor}{string, the political party of the author of the proposal. This is an
optional parameter and its default is empty.}

\item{siglaUfAutor}{string, the state (Unidade da Federacao) of the proposal author. This is an
optional parameter and its default is empty.}

\item{generoAutor}{string, the gender ("F" for female, "M" for male) of the proposal author. This
is an optional parameter and its default is empty.}

\item{idSituacaoProposicao}{string, the situation (stage of the legislative process) of the proposal
(check listarSituacoesProposicao for help). This is an optional parameter and its default is empty.}

\item{idOrgaoSituacaoProposicao}{string, the organ of the Camara dos Deputados where the proposition is
located (check listarTiposOrgaos function for help). This is an optional parameter and its default
is empty.}

\item{emTramitacao}{integer, indicates if the proposition (1) is still in the legislative process, or
(2) has reached an end. This is an optional parameter and its default is empty (both conditions return).}
}
\value{
A data frame containing detailed information on the requested proposals.
}
\description{
Returns a data frame containing detailed information on the requested proposals
from the Camara dos Deputados. sigla (type of proposal) and ano (year of the
proposal) are required parameters, unless parteNomeAutor (partial name of proposal author)
is filled in. Proposal name is always a combination of sigla (type of proposal),
numero (number of proposal) and ano (year of proposal).
}
\examples{

# Return a data frame containing all of the propositions written in 2015
proposicoes2015 <- listarProposicoes(sigla = "PL", ano = 2015)
str(proposicoes2015)

# Return a data frame containing the details of proposition PL-404/2015
proposicaoPL4042015 <- listarProposicoes(sigla = "PL", ano = 2015, numero = 404)
print(proposicaoPL4042015)

# Return a data frame containing all of the propositions written by legislator Luiza Erundina
proposicoesDepErundina <- listarProposicoes(parteNomeAutor = "Erundina")
str(proposicoesDepErundina)

# Return a data frame containing all of the propositions written in 2003 that are still in
# the legislative process
proposicoes2015EmTramitaco <- listarProposicoes(sigla = "PL", ano = 2015, emTramitacao = 1)
str(proposicoes2015EmTramitaco)

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

