% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listarProposicoesTramitadasNoPeriodo.R
\name{listarProposicoesTramitadasNoPeriodo}
\alias{listarProposicoesTramitadasNoPeriodo}
\title{List Propositions in Process}
\usage{
listarProposicoesTramitadasNoPeriodo(dtInicio, dtFim)
}
\arguments{
\item{dtInicio}{string of format dd/mm/yyyy, the initial date of for the requested
propositions.}

\item{dtFim}{string of format dd/mm/yyyy, the initial date of for the requested
propositions.}
}
\value{
A data frame containing the legislative actions taken by propositions.
}
\description{
Returns a data frame containing legislative actions taken
by proposition between and initial and a final date (parameters of the function) at
Camara dos Deputados. The maximum difference between  initial and final date
allowed by the web service allowed is 7 days. The inputs for this functions are of
class character and in the format dd/mm/yyyy.
}
\note{
The output of this function can be used as a parameter in functions that require
a vector of propositions.
}
\examples{
proposicoes <- listarProposicoesTramitadasNoPeriodo('01/08/2015', '08/08/2015')
head(proposicoes)
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

