% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listarProposicoesVotadasEmPlenario.R
\name{listarProposicoesVotadasEmPlenario}
\alias{listarProposicoesVotadasEmPlenario}
\title{List Voted Propositions}
\usage{
listarProposicoesVotadasEmPlenario(ano.param, tipo.param = "")
}
\arguments{
\item{ano}{integer, the year in which the propositions were submitted to the floor.}

\item{tipo}{character, the type of the proposition(s) (check listarSiglasTipoProposicao
function for help). If not specified the function will return all types of propositions.}
}
\value{
A data frame containing all the of propositions that were submitted to
the floor for voting.
}
\description{
Returns a data frame containing all the of propositions that were submitted to
the floor for voting at the Brazilian Chamber of Deputies in an specific year (ano) and
of a type (tipo). tipo is an optional parameter and if not specified the function will return
all types of propositions.
}
\note{
The output of this function can be used as a parameter in functions that require
voted propositions.
}
\examples{
# All the propositions voted in 2015
proposicoesVotadas2015 <- listarProposicoesVotadasEmPlenario(2015)
head(proposicoesVotadas2015)

# All the propositions of type "Proposta de Emenda à Constituição" (PEC, which
# are constitutional amendments) voted in 2015
pecsVotadas2015 <- listarProposicoesVotadasEmPlenario(2015, "PEC")
head(pecsVotadas2015)
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

