% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listarSiglasTipoProposicao.R
\name{listarSiglasTipoProposicao}
\alias{listarSiglasTipoProposicao}
\title{List Types of Proposals}
\usage{
listarSiglasTipoProposicao()
}
\value{
A data frame containing all the possible types of proposals.
}
\description{
Returns a data frame containing all the possible types of proposals (siglas)
at the Camara dos Deputados (for example, "PL" (Projeto de Lei), "MPV" (Medida
Provisoria, etc) and respective identification codes from the web service. This function
does not require any parameters.
}
\note{
The output of this function can be used as a parameter in functions that require
the type of proposals.
}
\examples{

siglas  <- listarSiglasTipoProposicao()
head(siglas)

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

