% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listarSituacoesProposicao.R
\name{listarSituacoesProposicao}
\alias{listarSituacoesProposicao}
\title{List Status of Propositions}
\usage{
listarSituacoesProposicao()
}
\value{
A data frame containing all the possible types of situations.
}
\description{
Returns a data frame containing all the possible status of propositions in
the legislative proccess at Camara dos Deputados (for example, "Vetado
totalmente" (entirely vetoed), Aguardando Parecer ("waiting for review"), etc) and
respectives identification codes at the web Service. This function does not require any
parameter.
}
\note{
The output of this function can be used as a parameter in functions that require
the type of status of a proposition.
}
\examples{
status  <- listarSituacoesProposicao()
head(status)
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

