% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listarSituacoesReuniaoSessao.R
\name{listarSituacoesReuniaoSessao}
\alias{listarSituacoesReuniaoSessao}
\title{List Status of Sessions}
\usage{
listarSituacoesReuniaoSessao()
}
\value{
A data frame containing all the status possibilities of sessions
}
\description{
Returns a data frame containing all the status possibilities of sessions at
Camara dos Deputados. This function does not require any parameters.
}
\note{
The output of this function can be used as a parameter in functions that require
the type of status of sessions.
}
\examples{

status  <- listarSituacoesReuniaoSessao()
head(status)

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

