% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listarTiposAutores.R
\name{listarTiposAutores}
\alias{listarTiposAutores}
\title{List types of Authors}
\usage{
listarTiposAutores()
}
\value{
A data frame containing all the possible types of authors of propositions.
}
\description{
Returns a data frame containing all the possible types of authors of propositions
at Camara dos Deputados (for example, "Bancada", "Deputado Federal", "Senador",
"Comissao Diretora", "Orgao do Poder Executivo", etc) and respectives identification
codes at the web service. This function does not require any parameter.
}
\note{
The output of this function can be used as a parameter in functions that require
the type of authors.
}
\examples{
autores <- listarTiposAutores()
head(autores)
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

