% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listarTiposOrgaos.R
\name{listarTiposOrgaos}
\alias{listarTiposOrgaos}
\title{List Camara dos Deputados Organizations}
\usage{
listarTiposOrgaos()
}
\value{
A data frame with information on Camara dos Deputados organizations.
}
\description{
Returns a data frame that lists internal Camara dos Deputados organizations
(for example, "Comissao", "Bancada") and respective identification codes from the web
service. This function does not require any parameters.
}
\note{
The output of this function can be used as a parameter in functions that require
the type of proposals.
}
\examples{

listarTiposOrgaos()

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

