% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obterAndamento.R
\name{obterAndamento}
\alias{obterAndamento}
\title{Get Proposal Progress}
\usage{
obterAndamento(sigla, numero, ano, dataIni = "", codOrgao = "")
}
\arguments{
\item{sigla}{string, the type of the proposal(s) (check listarSiglasTipoProposicao
function for help), which is part of the name of the proposal(s).}

\item{numero}{integer, the number of the proposal(s) (check listarSiglasTipoProposicao function
for help), which is part of the name of the proposal(s).}

\item{ano}{integer, the year of the proposal(s) (check listarSiglasTipoProposicao
function for help), which is part of the name of the proposal(s) and represents the year
the proposal was written.}

\item{dataIni}{string of format dd/mm/yyyy, the date of the sessions. This is an optional parameter
and its default is empty.}

\item{codOrgao}{integer, the Camara dos Deputados organ id code where the proposal is
located (check listarTiposOrgaos function for help). This is an optional parameter and its default
is empty.}
}
\value{
A data frame containing detailed information on the progress of the requested proposal.
}
\description{
Returns a data frame containing detailed information of the progress of the
requested proposal in the Brazilian Chamber of Deputies. sigla, numero and ano are required
parameters. Optionally, initial date (dataIni) and stance (codOrgao) can be specified. The proposal
name is always a combination of sigla (type of proposal), numero (number of proposal) and ano
(year of proposal).
}
\examples{

andamento <- obterAndamento("PL", 404, 2015)
head(andamento)

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

