% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obterDeputados.R
\name{obterDeputados}
\alias{obterDeputados}
\title{Get Legislator Info}
\usage{
obterDeputados()
}
\value{
A data frame containing basic information of all the active legislators in the Brazilian Chamber of Deputies.
}
\description{
Returns a data frame containing detailed information of the active legislators
at Camara dos Deputados and their respective information. There are none required parameters.
}
\examples{
# Return a data frame containing all of the active legislators
deputados <- obterDeputados()
print(deputados)
}
\author{
Alexia Aslan; Leonardo Sangali Barone;
}

