% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obterDetalhesDeputado.R
\name{obterDetalhesDeputado}
\alias{obterDetalhesDeputado}
\title{Get Legislator Details}
\usage{
obterDetalhesDeputado(ideCadastro, numLegislatura = "", atuacao = "bio")
}
\arguments{
\item{ideCadastro}{integer, the legislator id number from the web service (check obterDeputados
function for help).}

\item{numLegislatura}{integer, is the number of a Legislature. This is an optional parameter
and its default is empty.}

\item{atuacao}{string, if "bio" (default), return the basic bio information of the legislator;
if "comissoes", returns the committee information of the legislator; if "cargos", returns the
information on the positions occupied by the the legislator, if "exercicios" returns the time
intervals in which the legislator was at the Camara dos Deputados; if "filiacoes" returns the
political parties to which the legislator belonged; and if "lideranca", returns leadership
information on the legislators.}
}
\value{
A data frame containing detailed information on the legislator.
}
\description{
Returns a data frame containing detailed information on the legislator (for
example, political parties, leaderships, committee positions, etc) at Camara dos Deputados.
The parameter atuacao helps to define what type of information is going to be
returned.
}
\examples{

obterDetalhesDeputado(74784)
obterDetalhesDeputado(74784, atuacao = "cargos")

}
\author{
Alexia Aslan; Leonardo Sangali Barone;
}

