% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obterEmendasSubstitutivoRedacaoFinal.R
\name{obterEmendasSubstitutivoRedacaoFinal}
\alias{obterEmendasSubstitutivoRedacaoFinal}
\title{Get Proposition Amemndments, Substitutives and Final Draft}
\usage{
obterEmendasSubstitutivoRedacaoFinal(tipo, numero, ano)
}
\arguments{
\item{tipo}{string, the type of the proposition(s) (check listarSiglasTipoProposicao
function for help), which is part ot the name of the propostion(s).}

\item{numero}{integer, the number of the proposition(s) (check listarSiglasTipoProposicao function
for help), which is part ot the name of the propostion(s).}

\item{ano}{integer, the year of the proposition(s) (check listarSiglasTipoProposicao
function for help), which is part ot the name of the propostion(s) and represents the year
the proposition was written.}
}
\value{
A a data frame containing detailed information on the ammendments, substitutive
draft and final draft of the requested proposition.
}
\description{
Returns a data frame containing detailed information on the ammendments, substitutive
draft and final draft of the requested proposition in the Brazilian Chamber of Deputies.
sigla, numero and ano are required parameters. Proposition's name is always combination of tipo
(type of propostion), numero (number of proposition) and ano (year of propostion).
}
\examples{
# Proposition with amendments, substitutive and final draft
emendas <- obterEmendasSubstitutivoRedacaoFinal("PL", 3962, 2008)
head(emendas)

# Proposition without any amendments, substitutive and final draft
emendas <- obterEmendasSubstitutivoRedacaoFinal("PL", 404, 2015)
print(emendas)
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

