% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obterInteiroTeorDiscursosPlenario.R
\name{obterInteiroTeorDiscursosPlenario}
\alias{obterInteiroTeorDiscursosPlenario}
\title{Get Integral Speech}
\usage{
obterInteiroTeorDiscursosPlenario(codSessao, numOrador, numQuarto, numInsercao)
}
\arguments{
\item{codSessao}{string, the code of the session(check listarDiscursosPlenario function for help)}

\item{numOrador}{integer, the number of legislator in the session(check listarDiscursosPlenario
function for help)}

\item{numQuarto}{integer, the number of the room the speech is stored (check
listarDiscursosPlenario function for help)}

\item{numInsercao,}{the number of the speech in the session (check listarDiscursosPlenario
function for help)}
}
\value{
A a data frame containing detailed information on every speech given in
a legislative session.
}
\description{
Returns a data frame containing detailed information on every speech given in
a legislative session. All the parameter of the function are required. You can use the
listarDiscursosPlenario function to get values for the parameters.
}
\examples{
codSessao = "022.3.54.O"; numOrador=1 ;
numQuarto=11 ; numInsercao=0
discurso <- obterInteiroTeorDiscursosPlenario(codSessao, numOrador, numQuarto, numInsercao)
print(discurso)
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

