% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obterLideresBancadas.R
\name{obterLideresBancadas}
\alias{obterLideresBancadas}
\title{Get Leadership}
\usage{
obterLideresBancadas()
}
\value{
A data frame containing leaders and vice-leaders of parties/coalitions.
}
\description{
Returns a data frame containing leaders and vice-leaders of parties/coalitions at
Camara dos Deputados. There are no required parameters.
}
\examples{
# Return a data frame containing the learders and vice-leaders of parties/coalitions in office
lideres <- obterLideresBancadas()
print(lideres)
}
\author{
Alexia Aslan; Leonardo Sangali Barone;
}

