% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obterOrgaos.R
\name{obterOrgaos}
\alias{obterOrgaos}
\title{Get Camara dos Deputados Organizations}
\usage{
obterOrgaos()
}
\value{
A data frame with information on Camara dos Deputados organizations.
}
\description{
Returns a data frame that lists internal Camara dos Deputados organizations
(committees for example) and respective identification codes from the web
service. This function does not require any parameters.
}
\note{
The output of this function can be used as a parameter in functions that require
the type of proposals.
}
\examples{

obterOrgaos()

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

