% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obterPartidoCD.R
\name{obterPartidosCD}
\alias{obterPartidosCD}
\title{Get Parties}
\usage{
obterPartidosCD()
}
\value{
A data frame containing information on all the parties with representation.
}
\description{
Returns a data frame containing details of the parties with representation at
Camara dos deputados. There are no required parameters.
}
\examples{

# Return a data frame containing the parties with representation
partidos <- obterPartidosCD()
print(partidos)

}
\author{
Alexia Aslan; Leonardo Sangali Barone;
}

