% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obterPauta.R
\name{obterPauta}
\alias{obterPauta}
\title{Get Legislative Agenda}
\usage{
obterPauta(idOrgao, datIni, datFim)
}
\arguments{
\item{idOrgao}{integer, the Camara dos Deputados organ id code where the proposal is
located (check listarTiposOrgaos function for help).}

\item{datIni}{string of format dd/mm/yyyy, the initial date.}

\item{datFim}{string of format dd/mm/yyyy, the initial date.}
}
\value{
A data frame with information on a specific legislator attendance to sessions.
}
\description{
Returns a data frame with information on the legislative agenda of a Camara
dos Deputados organ between an initial and a final date. The maximum difference between
initial and final date allowed by the web service is 7 days. The inputs for this
function are of class character and in the format dd/mm/yyyy. All the three parameters
are required.
}
\examples{

# Members of Comissao de Agricultura, Pecuaria, Abastecimento e Desenvolvimento Rural
# between "10/04/2012" and "17/04/2012".
IDOrgao = 2004; dataInicial = "10/04/2012"; dataFinal = "17/04/2012"
pauta <- obterPauta(IDOrgao, dataInicial, dataFinal)
head(pauta)

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

