% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obterProposicao.R
\name{obterProposicao}
\alias{obterProposicao}
\title{Get Proposal Info}
\usage{
obterProposicao(tipo, numero, ano)
}
\arguments{
\item{tipo}{string, the type of the proposal(s) (check listarSiglasTipoProposicao
function for help), which is part of the name of the proposal(s).}

\item{numero}{integer, the number of the proposal(s) (check listarSiglasTipoProposicao function
for help), which is part of the name of the proposal(s).}

\item{ano}{integer, the year of the proposal(s) (check listarSiglasTipoProposicao
function for help), which is part of the name of the proposal(s) and represents the year
the proposal was written.}
}
\value{
A data frame containing detailed information on the requested proposals and related
proposals.
}
\description{
Returns a data frame containing detailed information the requested proposal
at Camara dos Deputados and respective related proposals ("proposicoes
apensadas"). All the parameters are required. Proposal name is always a combination of
tipo (type of proposal), numero (number of proposal) and ano (year of proposal). This
function is similar to the obterProposicaoPorID function.
}
\examples{

# Return a data frame containing all of proposition PL 404/2015
proposicao <- obterProposicao(tipo = "PL", numero = 404, ano = 2015)
print(proposicao)

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

