% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obterProposicaoPorID.R
\name{obterProposicaoPorID}
\alias{obterProposicaoPorID}
\title{Get Proposition Info by Id}
\usage{
obterProposicaoPorID(idProp)
}
\arguments{
\item{idProp}{integer, the requested proposal identification number.}
}
\value{
A data frame containing detailed information on the requested proposal and related
proposals.
}
\description{
Returns a data frame containing detailed information on the requested proposal
at Camara dos Deputados and respective related propositions ("proposicoes
apensadas"). The only parameter (idPorpo) is required. This function is similar to the
obterProposicao function.
}
\examples{

# Return a data frame containing all of proposal PL 404/2015, whose id is 947717
proposicao <- obterProposicaoPorID(idProp = 947717)
print(proposicao)

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

