% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obterVotacaoProposicao.R
\name{obterVotacaoProposicao}
\alias{obterVotacaoProposicao}
\title{Get Votes of a Proposition}
\usage{
obterVotacaoProposicao(tipo, numero, ano, bancada = F)
}
\arguments{
\item{tipo}{character, the type of the proposition(s) (check listarSiglasTipoProposicao
function for help), which is part ot the name of the propostion(s).}

\item{numero}{integer, the number of the proposition(s) (check listarSiglasTipoProposicao function
for help), which is part ot the name of the propostion(s).}

\item{ano}{integer, the year of the proposition(s) (check listarSiglasTipoProposicao
function for help), which is part ot the name of the propostion(s) and represents the year
the proposition was written.}

\item{bancada}{logic, if TRUE, leadership votes will be returned instead of
individual legislators vote.}
}
\value{
A data frame containing all the roll call votes for the requested
proposition and individual votes.
}
\description{
Returns a data frame containing all the roll call votes for the requested
proposition and individual votes at Camara dos Deputados. bancada is an
optional parameter and if bancada = TRUE, leadership votes will be returned instead of
individual legislators vote. Note that when a proposition goes to the floor every
amendment, substitutive propositions, etc, will be submitted to voting, not only the
main proposition (for example, one data frame may contain several different votings regarding
the requested proposition).
}
\examples{
# Return the roll call votes for the Proposta de Emenda a Constituicao 358 de 2013 (constitutional
# amendment 358 of 2013) voted in 2015
votesPec358.2013 <- obterVotacaoProposicao(tipo = "PEC", numero = 358,ano = 2013)
head(votesPec358.2013)

# Return the ledearship votes for the Proposta de Emenda a Constituicao 358 de 2013 (constitutional
# amendment 358 of 2013) voted in 2015
bancadaPec358.2013 <- obterVotacaoProposicao(tipo = "PEC", numero = 358,ano = 2013, bancada = TRUE)
head(bancadaPec358.2013)
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

