% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlAttributesToDataFrame.R
\name{xmlAttributesToDataFrame}
\alias{xmlAttributesToDataFrame}
\title{XML Attributes to Data Frame}
\usage{
xmlAttributesToDataFrame(doc, xpath)
}
\arguments{
\item{doc}{XMLInternalDocument, a xml document object.}

\item{xpath}{The xml path of the node set containing the attributes.}
}
\value{
A data frame containing attribute values
}
\description{
Returns the attribute values of a node set as a data frame.
}
\note{
This is an auxiliary function used in several other functions of the bRasilLeg package
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

