% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listarCargosOrgaosLegislativosCD.R
\name{listarCargosOrgaosLegislativosCD}
\alias{listarCargosOrgaosLegislativosCD}
\title{List Legislatives Positions at Camara dos Deputdaos}
\usage{
listarCargosOrgaosLegislativosCD()
}
\value{
A data frame with information on positions at Camara dos Deputados
}
\description{
Returns a data frame that lists positions (for example, "Presidente", "Relator",
etc) that can be occupied by legislators at an internal Camara dos Deputados organization
(comittes, floor, etc) and respectives identification codes at the web service. This function
does not require any parameter.
}
\note{
The output of this function can be used as a parameter in functions that require
the type of propositions.
}
\examples{
listarCargosOrgaosLegislativosCD()
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

