% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listarPresencasDia.R
\name{listarPresencasDia}
\alias{listarPresencasDia}
\title{List Legislator Attendance on one day}
\usage{
listarPresencasDia(data, numLegislatura = "", numMatriculaParlamentar = "",
  siglaPartido = "", siglaUF = "")
}
\arguments{
\item{data}{string of format dd/mm/yyyy, the date of the sessions.}

\item{numLegislatura}{integer, the number specific to the legislature.}

\item{numMatriculaParlamentar}{integer, the legislator "matricula" number (check the obterDeputados
function for help). This is an optional parameter and its default is empty.}

\item{siglaPartido}{string, the political party of the author of the proposal. This is an optional
parameter and its default is empty.}

\item{siglaUF}{string, the state (Unidade da Federacao) of the author of the proposal. This is an optional
parameter and its default is empty.}
}
\value{
A data frame with information on legislator attendance at sessions on a specific day.
}
\description{
Returns a data frame with information on legislator attendance at sessions
on a specific day. The maximum difference between initial and final date allowed by the
web service is 7 days. The inputs for this function are of class character and
in the format dd/mm/yyyy.
}
\examples{

dataSessao = "10/04/2012"
presencas <- listarPresencasDia(dataSessao)
head(presencas)

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

