% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listarPresencasParlamentar.R
\name{listarPresencasParlamentar}
\alias{listarPresencasParlamentar}
\title{List a Legislator Attendance}
\usage{
listarPresencasParlamentar(dataIni, dataFim, numMatriculaParlamentar)
}
\arguments{
\item{dataIni}{string of format dd/mm/yyyy, the initial date.}

\item{dataFim}{string of format dd/mm/yyyy, the initial date.}

\item{numMatriculaParlamentar}{integer, the legislator "matricula" number (check
obterDeputados function for help).}
}
\value{
A data frame with information on an specific legislator attendance to sessions.
}
\description{
Returns a data frame with information on an specific legislator attendance
to sessions between an initial and a final date. The maximum difference between initial
and final date allowed by the web service allowed is 7 days. The inputs for this
functions are of class character and in the format dd/mm/yyyy. All the three parameters
are required.
}
\examples{
dataInicial = "10/04/2012"; dataFinal = "17/04/2012"
presenca <- listarPresencasDia(dataInicial, dataFinal)
head(presencas)
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

