% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listarProposicoes.R
\name{listarProposicoes}
\alias{listarProposicoes}
\title{List Propositions}
\usage{
listarProposicoes(sigla = "", ano = "", numero = "",
  datApresentacaoIni = "", datApresentacaoFim = "", idTipoAutor = "",
  parteNomeAutor = "", siglaPartidoAutor = "", siglaUfAutor = "",
  generoAutor = "", idSituacaoProposicao = "",
  idOrgaoSituacaoProposicao = "", emTramitacao = "")
}
\arguments{
\item{sigla}{string, the type of the proposition(s) (check listarSiglasTipoProposicao
function for help), which is part ot the name of the propostion(s).}

\item{ano}{integer, the year of the proposition(s) (check listarSiglasTipoProposicao
function for help), which is part ot the name of the propostion(s) and represents the year
the proposition was written.}

\item{numero}{integer, the number of the proposition(s) (check listarSiglasTipoProposicao function
for help), which is part ot the name of the propostion(s). This is an optional parameter and
it's default is empty.}

\item{datApresentacaoIni}{string of format dd/mm/yyyy, the initial date of for the requested
propositions. This is a optional parameter and it's default is empty.}

\item{datApresentacaoFim}{string of format dd/mm/yyyy, the initial date of for the requested
propositions. This is a optional parameter and it's default is empty.}

\item{idTipoAutor}{character, the id code of the type of author of propositon (for example, "Bancada",
"Deputado Federal", "Senador", "Comissao Diretora", "Orgao do Poder Executivo", etc - check
listarTiposAutores function for help). This is a optional parameter and it's default is empty.}

\item{parteNomeAutor}{string, a part of the name of the propostion's author. This is a optional
parameter and it's default is empty.}

\item{siglaPartidoAutor}{string, the political part of the proposition's author. This is a optional
parameter and it's default is empty.}

\item{siglaUfAutor}{string, the state (Unidade da Federacao) of the proposition's author. This is
a optional parameter and it's default is empty.}

\item{generoAutor}{string, the gender ("F" for female "M" for male) of the proposition's author. This is
a optional parameter and it's default is empty.}

\item{idSituacaoProposicao}{string, the situation (stage of the legislative process) of the propostion
(check listarSituacoesProposicao for hel). This is a optional parameter and it's default is empty.}

\item{idOrgaoSituacaoProposicao}{string, the Camara dos Deputados stance where the proposition is
located (check listarTiposOrgaos function for help). This is a optional parameter and it's default
is empty.}

\item{emTramitacao}{integer, indicates if the proposition (1) is still in the legislative process, or
(2) has reached an end. This is a optional parameter and it's default is empty (both conditions return).}
}
\value{
A data frame containing detailed information the requested propositions.
}
\description{
Returns a data frame containing detailed information the requested propositions
at Camara dos Deputados. sigla (type of propostion) and ano (year of the
proposition) are required parameters, unless parteNomeAutor (author's proposition partial
name) is filled in. Proposition's name is always a combination of sigla (type of propostion),
numero (number of proposition) and ano (year of propostion).
}
\examples{
# Return a data frame containing all of the propositions written in 2015
proposicoes2015 <- listarProposicoes(sigla = "PL", ano = 2015)
str(proposicoes2015)

# Return a data frame containing the details of proposition PL-404/2015
proposicaoPL4042015 <- listarProposicoes(sigla = "PL", ano = 2015, numero = 404)
print(proposicaoPL4042015)

# Return a data frame containing all of the propositions written by legislator Luiza Erundina
proposicoesDepErundina <- listarProposicoes(parteNomeAutor = "Erundina")
str(proposicoesDepErundina)

# Return a data frame containing all of the propositions written in 2003 that are still in
# the legislative process
proposicoes2015EmTramitaco <- listarProposicoes(sigla = "PL", ano = 2015, emTramitacao = 1)
str(proposicoes2015EmTramitaco)
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

