% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listarProposicoesTramitadasNoPeriodo.R
\name{listarProposicoesTramitadasNoPeriodo}
\alias{listarProposicoesTramitadasNoPeriodo}
\title{List Proposals in Process}
\usage{
listarProposicoesTramitadasNoPeriodo(dtInicio, dtFim)
}
\arguments{
\item{dtInicio}{string of format dd/mm/yyyy, the initial date of the requested
proposals.}

\item{dtFim}{string of format dd/mm/yyyy, the initial date of the requested
proposals.}
}
\value{
A data frame containing the legislative actions regarding the proposals.
}
\description{
Returns a data frame containing legislative actions taken
regarding the proposal between an initial and a final date (parameters of the function) at
Camara dos Deputados. The maximum difference between the initial and final date
allowed by the web service is 7 days. The inputs for this function are of
class character and in the format dd/mm/yyyy.
}
\note{
The output of this function can be used as a parameter in functions that require
a vector of proposals.
}
\examples{

proposicoes <- listarProposicoesTramitadasNoPeriodo('01/08/2015', '08/08/2015')
head(proposicoes)

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

