% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listarProposicoesVotadasEmPlenario.R
\name{listarProposicoesVotadasEmPlenario}
\alias{listarProposicoesVotadasEmPlenario}
\title{List Proposals Voted On}
\usage{
listarProposicoesVotadasEmPlenario(ano, tipo = "")
}
\arguments{
\item{ano}{integer, the year in which the proposals were submitted to the floor.}

\item{tipo}{character, the type of the proposal(s) (check listarSiglasTipoProposicao
function for help). If not specified the function will return all types of proposals.}
}
\value{
A data frame containing all the of proposals that were submitted to
the floor for voting.
}
\description{
Returns a data frame containing all the of proposals that were submitted to
the floor for voting at Camara dos Deputados in a specific year (ano) and
of a type (tipo). tipo is an optional parameter and if not specified the function will return
all types of proposals.
}
\note{
The output of this function can be used as a parameter in functions that require
proposals already voted on.
}
\examples{

# All the propositions voted in 2015
proposicoesVotadas2015 <- listarProposicoesVotadasEmPlenario(2015)
head(proposicoesVotadas2015)

# All the propositions of type "Proposta de Emenda a Constituicao" (PEC, which
# are constitutional amendments) voted in 2015
pecsVotadas2015 <- listarProposicoesVotadasEmPlenario(2015, "PEC")
head(pecsVotadas2015)

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

