% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listarSiglasTipoProposicao.R
\name{listarSiglasTipoProposicao}
\alias{listarSiglasTipoProposicao}
\title{List Types of Propositions}
\usage{
listarSiglasTipoProposicao()
}
\value{
A data frame containing all the possible types of propositions.
}
\description{
Returns a data frame containing all the possible types of propositions (siglas)
in the Brazilian Chamber of Deputies (for example, "PL" (Projeto de Lei), "MPV" (Medida
Provisória, etc) and respectives identification codes at the web service. This function
does not require any parameter.
}
\note{
The output of this function can be used as a parameter in functions that require
the type of propositions.
}
\examples{
siglas  <- listarSiglasTipoProposicao()
head(siglas)
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

