% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listarSituacoesProposicao.R
\name{listarSituacoesProposicao}
\alias{listarSituacoesProposicao}
\title{List Status of Proposals}
\usage{
listarSituacoesProposicao()
}
\value{
A data frame containing all the possible status types.
}
\description{
Returns a data frame containing all status possibilities of proposals in
the legislative process at Camara dos Deputados (for example, "Vetado
totalmente" (entirely vetoed), Aguardando Parecer ("waiting for review"), etc) and
respective identification codes from the web Service. This function does not require any
parameters.
}
\note{
The output of this function can be used as a parameter in functions that require
the type of status of a proposal.
}
\examples{

status  <- listarSituacoesProposicao()
head(status)

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

