% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listarSituacoesReuniaoSessao.R
\name{listarSituacoesReuniaoSessao}
\alias{listarSituacoesReuniaoSessao}
\title{List Status of Sessions}
\usage{
listarSituacoesReuniaoSessao()
}
\value{
A data frame containing all the possible status of sessions
}
\description{
Returns a data frame containing all the possible status of sessions in
the Brazilian Chamber of Deputies. This function does not require any parameter.
}
\note{
The output of this function can be used as a parameter in functions that require
the type of status of session.
}
\examples{
status  <- listarSituacoesReuniaoSessao()
head(status)
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

