% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listarTiposAutores.R
\name{listarTiposAutores}
\alias{listarTiposAutores}
\title{List types of Authors}
\usage{
listarTiposAutores()
}
\value{
A data frame containing all the possible types of authors of proposals.
}
\description{
Returns a data frame containing all the possible types of authors of proposals
at the Camara dos Deputados (for example, "Bancada", "Deputado Federal", "Senador",
"Comissao Diretora", "Orgao do Poder Executivo", etc) and respective identification
codes from the web service. This function does not require any parameters.
}
\note{
The output of this function can be used as a parameter in functions that require
the type of authors.
}
\examples{

autores <- listarTiposAutores()
head(autores)

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

