% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listarTiposOrgaos.R
\name{listarTiposOrgaos}
\alias{listarTiposOrgaos}
\title{List C창mara dos Deputados Organizations}
\usage{
listarTiposOrgaos()
}
\value{
A data frame with information on C창mara dos Deputados organizations.
}
\description{
Returns a data frame that lists internal C창mara dos Deputados organizations
(for example, "Comiss찾o", "Bancada") and respectives identification codes at the web
service. This function does not require any parameter.
}
\note{
The output of this function can be used as a parameter in functions that require
the type of propositions.
}
\examples{
listarTiposOrgaos()
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

