% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obterAndamento.R
\name{obterAndamento}
\alias{obterAndamento}
\title{Get Proposition Progress}
\usage{
obterAndamento(sigla, numero, ano, dataIni = "", codOrgao = "")
}
\arguments{
\item{sigla}{string, the type of the proposition(s) (check listarSiglasTipoProposicao
function for help), which is part ot the name of the propostion(s).}

\item{numero}{integer, the number of the proposition(s) (check listarSiglasTipoProposicao function
for help), which is part ot the name of the propostion(s).}

\item{ano}{integer, the year of the proposition(s) (check listarSiglasTipoProposicao
function for help), which is part ot the name of the propostion(s) and represents the year
the proposition was written.}

\item{dataIni}{string of format dd/mm/yyyy, the date if the sessions. This is a optional parameter
and it's default is empty.}

\item{codOrgao}{integer, the Camara dos Deputados stance id code where the proposition is
located (check listarTiposOrgaos function for help). This is a optional parameter and it's default
is empty.}
}
\value{
A data frame containing detailed information on the progess of the requested proposition.
}
\description{
Returns a data frame containing detailed information of the progress of the
requested proposition in the Brazilian Chamber of Deputies. sigla, numero and ano are required
parameters. Optionally, initial date (dataIni) and stance (codOrgao) can be specified. Proposition's
name is always combination of sigla (type of propostion), numero (number of proposition) and ano
(year of propostion).
}
\examples{
andamento <- obterAndamento("PL", 404, 2015)
head(andamento)
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

