% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obterDeputados.R
\name{obterDeputados}
\alias{obterDeputados}
\title{Get Legislator Info}
\usage{
obterDeputados()
}
\value{
A data frame containing basic information on all the active legislators in the Brazilian Chamber of Deputies.
}
\description{
Returns a data frame containing detailed information on the active legislators
at Camara dos Deputados. There are no required parameters.
}
\examples{

# Return a data frame containing all of the active legislators
deputados <- obterDeputados()
print(deputados)

}
\author{
Alexia Aslan; Leonardo Sangali Barone;
}

