% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obterDetalhesDeputado.R
\name{obterDetalhesDeputado}
\alias{obterDetalhesDeputado}
\title{Get Legislator Details}
\usage{
obterDetalhesDeputado(ideCadastro, numLegislatura = "", atuacao = "bio")
}
\arguments{
\item{ideCadastro}{integer, the legislator id number at the web service (check obterDeputados
function for help).}

\item{numLegislatura}{integer, is the number of a Legislature. This is an optional parameter
and it's default is empty.}

\item{atuacao}{string, if "bio" (default), return the basic bio information of the legislator;
if "comissoes", return the committee information of the legislator; if "cargos", return the
information on the positions occupied by the the legilator, if "exercicios" return the time
intervals in which the legislator were at Câmara dos Deputados; if "filiacoes" return the
political parties in which the legislators belonged; and if "lideranca", returns leadership
information of the legislators.}
}
\value{
A data frame containing detailed information of the legislator.
}
\description{
Returns a data frame containing detailed information of the legislator (for
example, political parties, leaderships, committee positions, etc) in the Brazilian Chamber
of Deputies. The parameter atuacao helps to define what type of information is going to be
returned.
}
\examples{
obterDetalhesDeputado(74784)
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

