% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obterEmendasSubstitutivoRedacaoFinal.R
\name{obterEmendasSubstitutivoRedacaoFinal}
\alias{obterEmendasSubstitutivoRedacaoFinal}
\title{Get Proposal Amendments, Substitutions and Final Draft}
\usage{
obterEmendasSubstitutivoRedacaoFinal(tipo, numero, ano)
}
\arguments{
\item{tipo}{string, the type of the proposal(s) (check listarSiglasTipoProposicao
function for help), which is part of the name of the proposal(s).}

\item{numero}{integer, the number of the proposal(s) (check listarSiglasTipoProposicao function
for help), which is part o the name of the proposal(s).}

\item{ano}{integer, the year of the proposal(s) (check listarSiglasTipoProposicao
function for help), which is part of the name of the proposal(s) and represents the year
the proposal was written.}
}
\value{
A a data frame containing detailed information on the amendments, substitute
drafts and final drafts of the requested proposal.
}
\description{
Returns a data frame containing detailed information on the amendments, substitute
drafts and final drafts of the requested proposal at Camara dos Deputados.
sigla, numero and ano are required parameters. Proposal name is always a combination of tipo
(type of proposal), numero (number of proposal) and ano (year of proposal).
}
\examples{

# Proposition with amendments, substitutive and final draft
emendas <- obterEmendasSubstitutivoRedacaoFinal("PL", 3962, 2008)
head(emendas)

# Proposition without any amendments, substitutive and final draft
emendas <- obterEmendasSubstitutivoRedacaoFinal("PL", 404, 2015)
print(emendas)

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

