% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obterOrgaos.R
\name{obterOrgaos}
\alias{obterOrgaos}
\title{Get Camara dos Deputados Organizations}
\usage{
obterOrgaos()
}
\value{
A data frame with information on Camara dos Deputados organizations.
}
\description{
Returns a data frame that lists internal Camara dos Deputados organizations
(comittees for example) and respectives identification codes at the web
service. This function does not require any parameter.
}
\note{
The output of this function can be used as a parameter in functions that require
the type of propositions.
}
\examples{
obterOrgaos()
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

