% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obterProposicao.R
\name{obterProposicao}
\alias{obterProposicao}
\title{Get Proposition Info}
\usage{
obterProposicao(tipo, numero, ano)
}
\arguments{
\item{tipo}{string, the type of the proposition(s) (check listarSiglasTipoProposicao
function for help), which is part ot the name of the propostion(s).}

\item{numero}{integer, the number of the proposition(s) (check listarSiglasTipoProposicao function
for help), which is part ot the name of the propostion(s).}

\item{ano}{integer, the year of the proposition(s) (check listarSiglasTipoProposicao
function for help), which is part ot the name of the propostion(s) and represents the year
the proposition was written.}

\item{data}{string of format dd/mm/yyyy, the date if the sessions.}

\item{idOrgaoSituacaoProposicao}{string, the Camara dos Deputados stance where the proposition is
located (check listarTiposOrgaos function for help). This is a optional parameter and it's default
is empty.}
}
\value{
A data frame containing detailed information the requested propositions and attached
propositions.
}
\description{
Returns a data frame containing detailed information the requested proposition
in the Brazilian Chamber of Deputies and respectives attached propositions ("proposições
apensadas"). All the parameters are requires. Proposition's name is always a combination of
tipo (type of propostion), numero (number of proposition) and ano (year of propostion). This
function is similar to the obterProposicaoPorID function.
}
\examples{
# Return a data frame containing all of proposition PL 404/2015
proposicao <- obterProposicao(tipo = "PL", numero = 404, ano = 2015)
print(proposicao)
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

