% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obterProposicaoPorID.R
\name{obterProposicaoPorID}
\alias{obterProposicaoPorID}
\title{Get Proposition Info by Id}
\usage{
obterProposicaoPorID(idProp)
}
\arguments{
\item{numero}{integer, the requested proposition identification number.}
}
\value{
A data frame containing detailed information the requested propositions and attached
propositions.
}
\description{
Returns a data frame containing detailed information the requested proposition
in the Brazilian Chamber of Deputies and respectives attached propositions ("proposições
apensadas"). The only parameter (idPorpo) is requires. This function is similar to the
obterProposicao function.
}
\examples{
# Return a data frame containing all of proposition PL 404/2015, whose id is 947717
proposicao <- obterProposicaoPorID(idProp = 947717)
print(proposicao)
}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

