% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obterVotacaoProposicao.R
\name{obterVotacaoProposicao}
\alias{obterVotacaoProposicao}
\title{Get Votes on a Proposal}
\usage{
obterVotacaoProposicao(tipo, numero, ano, bancada = F)
}
\arguments{
\item{tipo}{character, the type of the proposal(s) (check listarSiglasTipoProposicao
function for help), which is part of the name of the proposal(s).}

\item{numero}{integer, the number of the proposal(s) (check listarSiglasTipoProposicao function
for help), which is part of the name of the proposal(s).}

\item{ano}{integer, the year of the proposal(s) (check listarSiglasTipoProposicao
function for help), which is part of the name of the proposal(s) and represents the year
the proposal was written.}

\item{bancada}{logic, if TRUE, leadership votes will be returned instead of
individual legislators' votes.}
}
\value{
A data frame containing all the roll call votes for the requested
proposal and individual votes.
}
\description{
Returns a data frame containing all the roll call votes for the requested
proposal and individual votes at Camara dos Deputados. bancada is an
optional parameter and if bancada = TRUE, leadership votes will be returned instead of
individual legislators' votes. Note that when a proposal goes to the floor every
amendment, substitute proposal etc, will be submitted for voting, not only the
main proposal (for example, one data frame may contain several different votes regarding
the requested proposal).
}
\examples{

# Return the roll call votes for the Proposta de Emenda a Constituicao 358 de 2013 (constitutional
# amendment 358 of 2013) voted in 2015
votesPec358.2013 <- obterVotacaoProposicao(tipo = "PEC", numero = 358,ano = 2013)
head(votesPec358.2013)

# Return the leadership votes for the Proposta de Emenda a Constituicao 358 de 2013 (constitutional
# amendment 358 of 2013) voted in 2015
bancadaPec358.2013 <- obterVotacaoProposicao(tipo = "PEC", numero = 358,ano = 2013, bancada = TRUE)
head(bancadaPec358.2013)

}
\author{
Leonardo Sangali Barone; Alexia Aslan
}

